+++ ft_memcpy.c
@@ -6,14 +6,16 @@
 /*   By: jan <marvin@42.fr>                         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/01 12:48:22 by jan               #+#    #+#             */
-/*   Updated: 2024/10/01 17:01:54 by jan              ###   ########.fr       */
+/*   Updated: 2024/10/06 22:36:24 by jans             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 /*
 #include <stdio.h>
 
-void	*ft_memcpy(void *dest, const void *src, unsigned int n);
+void	*ft_memcpy(void *dest, const void *src, size_t n);
 
 int	main()
 {
@@ -24,7 +26,7 @@ int	main()
 	return (0);
 }
 */
-void	*ft_memcpy(void *dest, const void *src, unsigned int n)
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned char		*d;
 	const unsigned char	*s;
